WIP: Remove vendored conversion_gen and use Kube 1.30 version
NOTE: This commit still needs work. Generated files were _not_ included.

Specifically, many of the CRD will need fixing because the updated version
of conversion_gen registers some of code comments as being something to
put into the YAML; the comments will need to be re-arranged in the Go files.

Signed-off-by: Nolan Brubaker <nolan@nbrubaker.com>